Continuación implementación de pruebas y seguimiento de gestión de calidad

+++ src/main/java/Lorem_Software/Library_Maintenance_System/business/controller/EjemplarController.java
@@ -3,7 +3,6 @@ package Lorem_Software.Library_Maintenance_System.business.controller;
 import java.util.Collections;
 import java.util.List;
 import java.util.ArrayList;
-import java.util.Optional;
 
 // import org.hibernate.mapping.List;
 import org.slf4j.Logger;
@@ -15,7 +14,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import jakarta.persistence.PreRemove;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 
 import Lorem_Software.Library_Maintenance_System.business.entity.Ejemplar;
@@ -34,7 +33,7 @@ public class EjemplarController {
     @Autowired
     private TituloDAO tituloDAO;
 
-    Ejemplar ejemplar = new  Ejemplar();
+    Ejemplar ejemplar = new Ejemplar();
 
     @GetMapping("/ListarEjemplares/{idTitulo}")
     public String listarEjemplares(@PathVariable("idTitulo") long idTitulo, Model model){
@@ -79,16 +78,23 @@ public class EjemplarController {
     }
 
     @PostMapping("/AltaEjemplar")
-    public String ejemplarSubmit(@ModelAttribute Ejemplar ejemplar, Model model) {
+    public String ejemplarSubmit(@ModelAttribute Ejemplar ejemplar, Model model, RedirectAttributes attribute) {
+        log.info("El id del ejemplar es"+ejemplar.getIdEjemplar());
         this.ejemplar = ejemplar;
         model.addAttribute("ejemplar", ejemplar);
         ejemplarDAO.save(ejemplar);
+        attribute.addFlashAttribute("success", "El ejemplar se ha dado correctamente de alta");
         return "redirect:/ListarTitulos";
     }
 
     @GetMapping("/AltaEjemplar/delete/{IdPrestamo}")
-	public String eliminarTitulo(@PathVariable("IdPrestamo") long IdPrestamo){
+	public String eliminarTitulo(@PathVariable("IdPrestamo") long IdPrestamo, RedirectAttributes attribute){
+        if(ejemplarDAO.findById(IdPrestamo).get().getPrestamo()!=null || ejemplarDAO.findById(IdPrestamo).get().getReserva() != null){
+            attribute.addFlashAttribute("error", "No se puede eliminar el ejemplar porque está prestado o tiene una reserva");
+            return "redirect:/ListarEjemplares/"+ejemplarDAO.findById(IdPrestamo).get().getTit().getId();
+        }
 		ejemplarDAO.deleteById(IdPrestamo);
+        attribute.addFlashAttribute("warning", "El ejemplar ha sido eliminado");
 		return "redirect:/ListarTitulos";
 	}
